+++ NEWS.md
@@ -4,6 +4,7 @@
 
 -   The `prep_data()`, `hbam()`, and `fbam()` functions now allow users to not supply self-placements. In this case, no meaningful respondent positions will be estimated, but all other parameters are unaffected.
 -   The `prep_data()` function now allows missing values in the `group_id` vector and will drop respondents who do not have a valid `group_id`. Missing values would previously generate an uninformative error message. 
+-   The models now accept even numbers of answering categories, but `prep_data()` will throw a warning.
 
 # hbamr 2.1.2